requirements.txt
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.20
cmake==3.26.3
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
diffusers==0.15.0
duckdb==0.7.1
entrypoints==0.4
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gptcache==0.1.10
gradio==3.27.0
gradio_client==0.1.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.13.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.139
linkify-it-py==2.0.0
lit==16.0.1
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.4
openapi-schema-pydantic==1.2.4
orjson==3.8.10
packaging==23.1
pandas==1.5.3
parsel==1.7.0
Pillow==9.5.0
posthog==2.5.0
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.15.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
queuelib==1.6.2
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
streamlit==1.21.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tldextract==3.4.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.28.0
triton==2.0.0
Twisted==22.10.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
w3lib==2.1.1
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.1
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
zstandard==0.20.0